refactor(sdk): replace privacy pool with mock pool contract
- Create mock-pool-contract.ts with contract-compatible async interface
- Add mock-server-action.ts type for callback pattern
- Update transfers, discovery, helpers, contracts to use new types
- Remove old pool.ts
