+++ sdk/src/testing/contracts.ts
@@ -6,7 +6,7 @@
 import type { Amount, NoteId, StarknetAddress } from "../interfaces.js";
 import { AddressMap, toBigInt } from "../utils/index.js";
 import { assert } from "../utils/validation.js";
-import type { PrivacyPool } from "./pool.js";
+import type { MockPoolContract } from "./mock-pool-contract.js";
 
 /** Interface for any mock contract */
 export interface MockContract {
@@ -92,7 +92,7 @@ export class MockSwapHelper implements MockContract {
     this.contracts.get(fromToken).setBalance(this.address, 0n);
     this.contracts.get(toToken).setBalance(this.address, amount * 2n);
     this.contracts
-      .get<PrivacyPool>(toBigInt(poolAddress))
+      .get<MockPoolContract>(toBigInt(poolAddress))
       .openDeposit(toBigInt(noteId), toBigInt(toToken), amount * 2n);
   }
 }
